Mount /var/www/html as one, not individual subdirectories
Mounting individual sub-directories (config, data, etc.) is nice and clean,
but doesn't work well, because Nextcloud wants to write files to
`/var/www/html`.

Most importantly, `version.php`, which is later used on every container
startup to compare versions and decide whether an upgrade is necessary.
If we don't persist the `version.php` file across container restarts
(which used to be the case until now), on each startup, Nextcloud would
log an attempt to do an initial installation. This couldn't go very far,
but still..

Another important file that we weren't persisting is
`/var/www/html/.htaccess`, which was causing pretty URLs to break.
The `.htaccess` file is generated correctly after an initial installation
or with a manual call to `php /var/www/html/occ maintenance:update:htaccess`.