add fail2ban usage
Install fail2ban and setup its usage.
Configuration is based on https://docs.nextcloud.com/server/21/admin_manual/installation/harden_server.html#setup-fail2ban
The jail variables are mostly configurable and also an enable value is present,
however it only controls if the service is enabled/disabled does not uninstall the package.
This is mostly due to laziness but as this is the advised usage, I leave it like this for now.
Port variable is also configurable, hopefully it is enough to use the playbook with
reverse-proxy as well.